# [SPARK-26572][SQL] fix aggregate codegen result evaluation