@maropu 2019-02-04
I think we should handle this case in a planner?
For example, if we turn off broadcast join, the behaviour changes;
```
scala> val baseTable = Seq((1), (1)).toDF("idx")
scala> val distinctWithId = baseTable.distinct.withColumn("id", functions.monotonically_increasing_id())
scala> baseTable.join(distinctWithId, "idx").show
+---+------------+
|idx|          id|
+---+------------+
|  1|369367187456|
|  1|369367187457|
+---+------------+

sql("SET spark.sql.autoBroadcastJoinThreshold=-1")
scala> baseTable.join(distinctWithId, "idx").show
+---+------------+
|idx|          id|
+---+------------+
|  1|369367187456|
|  1|369367187456|
+---+------------+
```
Could you check again?