@maropu 2019-02-04
But, the current fix changes the change behaivour of the other stateful exprs? e.g.,
```
scala> val baseTable = Seq((1), (1)).toDF("idx")
scala> val distinctWithRandId = baseTable.distinct.withColumn("id", functions.rand())
scala> baseTable.join(distinctWithRandId, "idx").show
// w/ this pr
+---+-------------------+                                                       
|idx|                 id|
+---+-------------------+
|  1|0.09884607470357032|
|  1|0.09884607470357032|
+---+-------------------+

// the current master
scala> baseTable.join(distinctWithRandId, "idx").show
+---+-------------------+                                                       
|idx|                 id|
+---+-------------------+
|  1|0.08661806632196123|
|  1| 0.0903755555098934|
+---+-------------------+
```