@peter-toth 2019-02-04
Isn't the first one (with same id) the expected result here?
I mean, if you run it on master with `spark.sql.codegen.wholeStage=false` you get same ids.