@maropu 2019-02-04
This issue happens in case of stateful exprs only? If so, could you modify the code to apply the current fix only if `HashAggregateExec` has stateful exprs? I worry about the performance regression @mgaido91 pointed out, too. It seems the current fix affect the other queries, its a corner case though....