@peter-toth 2019-02-04
Here are my benchmark results of q55. I run 3 times on master and 3 times on this PR branch against `scale=5` generated data.
Master:
```
master:
  Stopped after 5 iterations, 29324 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_162-b12 on Mac OS X 10.14.2
Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
TPCDS Snappy:                            Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
q55                                           5683 / 5865          2.6         391.2       1.0X

  Stopped after 5 iterations, 28914 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_162-b12 on Mac OS X 10.14.2
Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
TPCDS Snappy:                            Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
q55                                           5584 / 5783          2.6         384.3       1.0X

  Stopped after 5 iterations, 29905 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_162-b12 on Mac OS X 10.14.2
Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
TPCDS Snappy:                            Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
q55                                           5873 / 5981          2.5         404.3       1.0X
```
This PR:
```
this PR:
  Stopped after 5 iterations, 32577 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_162-b12 on Mac OS X 10.14.2
Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
TPCDS Snappy:                            Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
q55                                           6226 / 6515          2.3         428.5       1.0X


  Stopped after 5 iterations, 30612 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_162-b12 on Mac OS X 10.14.2
Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
TPCDS Snappy:                            Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
q55                                           5792 / 6122          2.5         398.6       1.0X

  Stopped after 5 iterations, 32918 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_162-b12 on Mac OS X 10.14.2
Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
TPCDS Snappy:                            Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
q55                                           6415 / 6584          2.3         441.5       1.0X
```
Although the results are a bit varying, it seems this patch would introduce some performance degradation.
I will try to modify the patch to evaluate only `Stateful` expressions as @maropu suggested and run the benchmark again.